+++ ML-Frameworks/tensorflow-aarch64/get-source.sh
@@ -21,7 +21,7 @@ source ../utils/git-utils.sh
 
 set -eux -o pipefail
 
-TENSORFLOW_HASH=ab8aab720f1648f6a470b159b0d1aea3a5b0df81 # 2.20.0-dev0 from master, 25th July 2025
+TENSORFLOW_HASH=6aa8fd07270293b918255af5988aef45b844c5b4 # from nightly, September 22nd
 
 git-shallow-clone https://github.com/tensorflow/tensorflow.git $TENSORFLOW_HASH
 
@@ -30,6 +30,6 @@ git-shallow-clone https://github.com/tensorflow/tensorflow.git $TENSORFLOW_HASH
 
     # Apply TensorFlow WIP patches here
 
-    # https://github.com/tensorflow/tensorflow/pull/93951 - build(aarch64): Update to oneDNN-3.7 + ACL-24.12 (fix)
-    apply-github-patch tensorflow/tensorflow e9cd9ea7d5f52b6c6740ce4e486e0fce554a26a4
+    # https://github.com/tensorflow/tensorflow/pull/100882 - build(aarch64): Update Compute Library to 52.4.0
+    apply-github-patch tensorflow/tensorflow 0cd9bed2d5f5456576d3bcd0b80aba7cd478ee26
 )